refactoring of Synthesis/Installment IU hierarchy: 	- tree structured output is now pushed to the pattern-demo directory (as it's used only there) 	- PhraseBasedIncrInstallmentIU now directly inherits from SysInstallmentIU 	- hesitation capabilities moved to car chase domain (it's only reasonably used there for now) 	- turn completion specific stuff moved to CompletionInstallmentIU
renamed "incremental.listener" to "incremental.sink" and moved our sources to "incremental.source" to reflect the fact that these IUmodules only use one right/left buffer

moved around code to reduce dependencies between packages in a lot of places

reduced the amount of Eclipse and Javadoc warnings in a lot of places 

git-svn-id: svn+ssh://gate.spectrum.uni-bielefeld.de/homes/tbaumann/inpro-svnrepo/trunk@2807 19460358-522e-0410-942e-fdb976ae3ca2

+++ src/inpro/pitch/PitchTracker.java
@@ -315,7 +315,7 @@ public class PitchTracker extends BaseDataProcessor {
 		//System.err.println("Setting minLag to " + minLag);
 		signalBuffer = new double[maxLag * 2 + 2];
 		Arrays.fill(signalBuffer, 0);
-		int queueSize = ((int) (maxLag / 160)) + 1; // should be 1 for freq_low > 100Hz, 2 for f_l > 50Hz, ...
+		int queueSize = (maxLag / 160) + 1; // should be 1 for freq_low > 100Hz, 2 for f_l > 50Hz, ...
 		pitchAssignmentQueue = new ArrayBlockingQueue<PitchedDoubleData>(queueSize, false);
 		listeners = ps.getComponentList(PROP_LISTENERS, SignalFeatureListener.class);
 	}
